+++ Assets/Nova/Resources/Shaders/FlipGrid.shader
@@ -67,7 +67,7 @@ Shader "Nova/VFX/Flip Grid"
                 float flipStartTime = gridDiagIndex * i.gridInfo.y;
                 float flipEndTime = flipStartTime + _FlipDuration;
 
-                fixed4 col = _BackColor;
+                float4 col = _BackColor;
 
                 if (flipStartTime > _T)
                 {
